requirements.txt
pypubsub
opencv-python